backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
feedparser==6.0.11
idna==3.6
joblib==1.3.2
newsapi-python==0.2.7
nltk==3.8.1
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
sgmllib3k==1.0.0
soupsieve==2.5
sqlparse==0.4.4
textblob==0.17.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
whitenoise==6.6.0
